プロジェクトのuseEventServiceAsProduct設定を適用

+++ dst/routes/offerCatalogs.js
@@ -31,6 +31,7 @@ exports.offerCatalogsRouter = offerCatalogsRouter;
 offerCatalogsRouter.all('/add', validateCsrfToken_1.validateCsrfToken, ...validate(true), 
 // tslint:disable-next-line:cyclomatic-complexity max-func-body-length
 (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+    var _a;
     try {
         const offerService = new sdk_1.chevre.service.Offer({
             endpoint: process.env.API_ENDPOINT,
@@ -52,6 +53,13 @@ offerCatalogsRouter.all('/add', validateCsrfToken_1.validateCsrfToken, ...valida
             auth: req.user.authClient,
             project: { id: req.project.id }
         });
+        const projectService = new sdk_1.chevre.service.Project({
+            endpoint: process.env.API_ENDPOINT,
+            auth: req.user.authClient,
+            project: { id: '' }
+        });
+        const chevreProject = yield projectService.findById({ id: req.project.id });
+        const useEventServiceAsProduct = ((_a = chevreProject.subscription) === null || _a === void 0 ? void 0 : _a.useEventServiceAsProduct) === true;
         let message = '';
         let errors = {};
         let csrfToken;
@@ -64,8 +72,10 @@ offerCatalogsRouter.all('/add', validateCsrfToken_1.validateCsrfToken, ...valida
                     req.body.id = '';
                     const { offerCatalogFromBody, serviceTypeFromBody } = yield createFromBody(req);
                     const offerCatalog = yield offerCatalogService.create(offerCatalogFromBody);
-                    // EventServiceプロダクトも作成
-                    yield upsertEventService(offerCatalog, serviceTypeFromBody)({ product: productService });
+                    if (!useEventServiceAsProduct) {
+                        // EventServiceプロダクトも作成
+                        yield upsertEventService(offerCatalog, serviceTypeFromBody)({ product: productService });
+                    }
                     // tslint:disable-next-line:no-dynamic-delete
                     delete req.session.csrfSecret;
                     req.flash('message', '登録しました');
@@ -111,6 +121,8 @@ offerCatalogsRouter.all('/add', validateCsrfToken_1.validateCsrfToken, ...valida
         let offers = [];
         if (Array.isArray(forms.itemListElement) && forms.itemListElement.length > 0) {
             const itemListElementIds = forms.itemListElement.map((element) => element.id);
+            // tslint:disable-next-line:no-suspicious-comment
+            // TODO カタログのアイテムリスト上限数への依存を排除
             const searchOffersResult = yield offerService.search({
                 limit: 100,
                 project: { id: { $eq: req.project.id } },
@@ -131,7 +143,8 @@ offerCatalogsRouter.all('/add', validateCsrfToken_1.validateCsrfToken, ...valida
             serviceTypes: searchServiceTypesResult.data,
             offers: offers,
             productTypes: productType_1.productTypes,
-            originalOfferCatalog
+            originalOfferCatalog,
+            useEventServiceAsProduct
         });
     }
     catch (error) {
@@ -151,7 +164,7 @@ function createCopiedString(params) {
 offerCatalogsRouter.all('/:id/update', ...validate(false), 
 // tslint:disable-next-line:max-func-body-length
 (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
-    var _a;
+    var _b, _c;
     try {
         const offerService = new sdk_1.chevre.service.Offer({
             endpoint: process.env.API_ENDPOINT,
@@ -173,11 +186,18 @@ offerCatalogsRouter.all('/:id/update', ...validate(false),
             auth: req.user.authClient,
             project: { id: req.project.id }
         });
+        const projectService = new sdk_1.chevre.service.Project({
+            endpoint: process.env.API_ENDPOINT,
+            auth: req.user.authClient,
+            project: { id: '' }
+        });
         const searchServiceTypesResult = yield categoryCodeService.search({
             limit: 100,
             project: { id: { $eq: req.project.id } },
             inCodeSet: { identifier: { $eq: sdk_1.chevre.factory.categoryCode.CategorySetIdentifier.ServiceType } }
         });
+        const chevreProject = yield projectService.findById({ id: req.project.id });
+        const useEventServiceAsProduct = ((_b = chevreProject.subscription) === null || _b === void 0 ? void 0 : _b.useEventServiceAsProduct) === true;
         const offerCatalog = yield offerCatalogService.findById({ id: req.params.id });
         const searchEventServicesResult = yield productService.search({
             limit: 1,
@@ -200,8 +220,10 @@ offerCatalogsRouter.all('/:id/update', ...validate(false),
                     req.body.id = req.params.id;
                     const { offerCatalogFromBody, serviceTypeFromBody } = yield createFromBody(req);
                     yield offerCatalogService.update(offerCatalogFromBody);
-                    // EventServiceプロダクトも編集(なければ作成)
-                    yield upsertEventService(offerCatalogFromBody, serviceTypeFromBody)({ product: productService });
+                    if (!useEventServiceAsProduct) {
+                        // EventServiceプロダクトも編集(なければ作成)
+                        yield upsertEventService(offerCatalogFromBody, serviceTypeFromBody)({ product: productService });
+                    }
                     req.flash('message', '更新しました');
                     res.redirect(req.originalUrl);
                     return;
@@ -214,7 +236,7 @@ offerCatalogsRouter.all('/:id/update', ...validate(false),
         const forms = Object.assign(Object.assign(Object.assign({ additionalProperty: [] }, offerCatalog), { 
             // 興行から興行区分を参照する(2022-09-03~)
             // serviceType: offerCatalog.itemOffered.serviceType?.codeValue,
-            serviceType: (_a = eventServiceProduct.serviceType) === null || _a === void 0 ? void 0 : _a.codeValue }), req.body);
+            serviceType: (_c = eventServiceProduct.serviceType) === null || _c === void 0 ? void 0 : _c.codeValue }), req.body);
         if (forms.additionalProperty.length < NUM_ADDITIONAL_PROPERTY) {
             // tslint:disable-next-line:prefer-array-literal
             forms.additionalProperty.push(...[...Array(NUM_ADDITIONAL_PROPERTY - forms.additionalProperty.length)].map(() => {
@@ -225,6 +247,8 @@ offerCatalogsRouter.all('/:id/update', ...validate(false),
         let offers = [];
         if (Array.isArray(forms.itemListElement) && forms.itemListElement.length > 0) {
             const itemListElementIds = forms.itemListElement.map((element) => element.id);
+            // tslint:disable-next-line:no-suspicious-comment
+            // TODO カタログのアイテムリスト上限数への依存を排除
             const searchOffersResult = yield offerService.search({
                 limit: 100,
                 project: { id: { $eq: req.project.id } },
@@ -241,7 +265,8 @@ offerCatalogsRouter.all('/:id/update', ...validate(false),
             offers: offers,
             forms: forms,
             serviceTypes: searchServiceTypesResult.data,
-            productTypes: productType_1.productTypes
+            productTypes: productType_1.productTypes,
+            useEventServiceAsProduct
         });
     }
     catch (error) {
@@ -289,6 +314,7 @@ offerCatalogsRouter.delete('/:id', (req, res) => __awaiter(void 0, void 0, void 
     }
 }));
 function preDelete(req, offerCatalog) {
+    var _a;
     return __awaiter(this, void 0, void 0, function* () {
         const eventService = new sdk_1.chevre.service.Event({
             endpoint: process.env.API_ENDPOINT,
@@ -300,19 +326,27 @@ function preDelete(req, offerCatalog) {
             auth: req.user.authClient,
             project: { id: req.project.id }
         });
+        const projectService = new sdk_1.chevre.service.Project({
+            endpoint: process.env.API_ENDPOINT,
+            auth: req.user.authClient,
+            project: { id: '' }
+        });
+        const chevreProject = yield projectService.findById({ id: req.project.id });
+        const useEventServiceAsProduct = ((_a = chevreProject.subscription) === null || _a === void 0 ? void 0 : _a.useEventServiceAsProduct) === true;
         // プロダクト確認
         if (offerCatalog.itemOffered.typeOf === sdk_1.chevre.factory.product.ProductType.EventService) {
-            // EventServiceについてはどうするか
-            // プロダクトのpreDelete後にEventServiceも削除
-            const searchEventServicesResult = yield productService.search({
-                limit: 1,
-                typeOf: { $eq: sdk_1.factory.product.ProductType.EventService },
-                productID: { $eq: `${sdk_1.factory.product.ProductType.EventService}${offerCatalog.id}` }
-            });
-            const existingEventService = searchEventServicesResult.data.shift();
-            if (existingEventService !== undefined) {
-                yield (0, products_1.preDelete)(req, existingEventService);
-                yield productService.deleteById({ id: String(existingEventService.id) });
+            if (!useEventServiceAsProduct) {
+                // プロダクトのpreDelete後にEventServiceも削除
+                const searchEventServicesResult = yield productService.search({
+                    limit: 1,
+                    typeOf: { $eq: sdk_1.factory.product.ProductType.EventService },
+                    productID: { $eq: `${sdk_1.factory.product.ProductType.EventService}${offerCatalog.id}` }
+                });
+                const existingEventService = searchEventServicesResult.data.shift();
+                if (existingEventService !== undefined) {
+                    yield (0, products_1.preDelete)(req, existingEventService);
+                    yield productService.deleteById({ id: String(existingEventService.id) });
+                }
             }
         }
         const searchProductsResult = yield productService.search({
@@ -362,6 +396,8 @@ offerCatalogsRouter.get('/:id/offers', (req, res) => __awaiter(void 0, void 0, v
         const limit = 100;
         const page = 1;
         let data;
+        // tslint:disable-next-line:no-suspicious-comment
+        // TODO カタログのアイテムリスト上限数への依存を排除
         const searchResult = yield offerService.search({
             limit: limit,
             page: page,
@@ -395,7 +431,7 @@ offerCatalogsRouter.get('', (__, res) => __awaiter(void 0, void 0, void 0, funct
     });
 }));
 offerCatalogsRouter.get('/getlist', (req, res) => __awaiter(void 0, void 0, void 0, function* () {
-    var _b, _c, _d, _e, _f, _g, _h, _j, _k;
+    var _d, _e, _f, _g, _h, _j, _k, _l, _m;
     try {
         const offerCatalogService = new sdk_1.chevre.service.OfferCatalog({
             endpoint: process.env.API_ENDPOINT,
@@ -421,15 +457,15 @@ offerCatalogsRouter.get('/getlist', (req, res) => __awaiter(void 0, void 0, void
             itemOffered: {
                 serviceType: {
                     codeValue: {
-                        $eq: (typeof ((_d = (_c = (_b = req.query.itemOffered) === null || _b === void 0 ? void 0 : _b.serviceType) === null || _c === void 0 ? void 0 : _c.codeValue) === null || _d === void 0 ? void 0 : _d.$eq) === 'string'
+                        $eq: (typeof ((_f = (_e = (_d = req.query.itemOffered) === null || _d === void 0 ? void 0 : _d.serviceType) === null || _e === void 0 ? void 0 : _e.codeValue) === null || _f === void 0 ? void 0 : _f.$eq) === 'string'
                             && req.query.itemOffered.serviceType.codeValue.$eq.length > 0)
                             ? req.query.itemOffered.serviceType.codeValue.$eq
                             : undefined
                     }
                 },
                 typeOf: {
-                    $eq: (typeof ((_f = (_e = req.query.itemOffered) === null || _e === void 0 ? void 0 : _e.typeOf) === null || _f === void 0 ? void 0 : _f.$eq) === 'string' && ((_h = (_g = req.query.itemOffered) === null || _g === void 0 ? void 0 : _g.typeOf) === null || _h === void 0 ? void 0 : _h.$eq.length) > 0)
-                        ? (_k = (_j = req.query.itemOffered) === null || _j === void 0 ? void 0 : _j.typeOf) === null || _k === void 0 ? void 0 : _k.$eq
+                    $eq: (typeof ((_h = (_g = req.query.itemOffered) === null || _g === void 0 ? void 0 : _g.typeOf) === null || _h === void 0 ? void 0 : _h.$eq) === 'string' && ((_k = (_j = req.query.itemOffered) === null || _j === void 0 ? void 0 : _j.typeOf) === null || _k === void 0 ? void 0 : _k.$eq.length) > 0)
+                        ? (_m = (_l = req.query.itemOffered) === null || _l === void 0 ? void 0 : _l.typeOf) === null || _m === void 0 ? void 0 : _m.$eq
                         : undefined
                 }
             }
@@ -454,7 +490,7 @@ offerCatalogsRouter.get('/getlist', (req, res) => __awaiter(void 0, void 0, void
     }
 }));
 offerCatalogsRouter.get('/searchOffersByPrice', (req, res) => __awaiter(void 0, void 0, void 0, function* () {
-    var _l;
+    var _o;
     try {
         const offerService = new sdk_1.chevre.service.Offer({
             endpoint: process.env.API_ENDPOINT,
@@ -471,7 +507,7 @@ offerCatalogsRouter.get('/searchOffersByPrice', (req, res) => __awaiter(void 0, 
                 'priceSpecification.price': sdk_1.chevre.factory.sortType.Descending
             },
             project: { id: { $eq: req.project.id } },
-            itemOffered: { typeOf: { $eq: (_l = req.query.itemOffered) === null || _l === void 0 ? void 0 : _l.typeOf } },
+            itemOffered: { typeOf: { $eq: (_o = req.query.itemOffered) === null || _o === void 0 ? void 0 : _o.typeOf } },
             priceSpecification: {
                 // 売上金額で検索
                 accounting: {
@@ -510,6 +546,8 @@ function createFromBody(req) {
                 };
             });
         }
+        // tslint:disable-next-line:no-suspicious-comment
+        // TODO カタログのアイテムリスト上限数への依存を排除
         const MAX_NUM_OFFER = 100;
         if (itemListElement.length > MAX_NUM_OFFER) {
             throw new Error(`オファー数の上限は${MAX_NUM_OFFER}です`);
